Fix for LPP-2181. Fix submitted by Michael Gregor
Test case - components/incubator/test/scrolledittext-test.lzx 

Type characters until they wrap... Notice that the characters go behind the scrollbar. The textfield width does not respond to the scrollbar vis. 

This component was updated with a dynamically instantiated scrollbar. This created a regression in the textfields behavior that would set the width upon the scrollbars visibility. This is because the constraint was delclared in the markup and the object does not exist. A delegate fixes the problem. 


git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@1101 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c