Change change.XXXXXXXXX.4kLDyQqX.txt by adam@axe.local /Users/adam/laszlo/src/svn/openlaszlo/trunk/ on 2006-09-13 16:30:54 PDT
Summary: Fix specificity implementation for CSS rules. Implement new "applicability" logic without resorting to poking at private aspects of the class implementation.

New Features:

Bugs Fixed:

Technical Reviewer: ptw
QA Reviewer: (pending)
Doc Reviewer: (pending)

Documentation:

Release Notes:
In cases where two rules have the same specificity, Laszlo defines an additional "applicability" rule, which attempts to prefer rules which select subclasses over rules which select superclasses.

In the case of descendant selectors, descendants are compared in lock-step. If two steps in the descendant selector select for non-equivalent classes, the subclass will be considered higher specificity than the superclass. 

In cases where one of the two descendants being compared is not a class selector, the selector is skipped for the purposes of this comparison. If the comparison process walks off the end of either selector chain, last-one-wins lexical order applies.

Details:
This is a refactor of the specificity rule calculation in addition to a removal of the skanky classhops count.

Tests:
   New test:
    http://localhost:8080/trunk/test/style/specificity/newspecificity.lzx

   Updated (to show how redefined applicability works) :
    http://localhost:8080/trunk/test/style/specificity/applicability.lzx

   Ran:
    http://localhost:8080/trunk/diamond/client/test/speedtest.lzx
    http://localhost:8080/trunk/test/style/descendantselector/complex-descendant-selector-test.lzx
    http://localhost:8080/trunk/test/style/constraints/constraint-test.lzx
    http://localhost:8080/trunk/test/style/specificity/specificity-tests.lzx

git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@1840 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c