Change 20070318-sallen-2 by sallen@sallen-new on 2007-03-18 09:41:16 PDT     in /cygdrive/c/laszlo/svn/openlaszlo/trunk/WEB-INF/lps/lfc/views     for http://svn.openlaszlo.org/openlaszlo/trunk/WEB-INF/lps/lfc/views
Summary: default value for clickable attribute of view

Bugs Fixed: LPP-3734 checking clickable attribute of a view should not generate a debugger warning

Technical Reviewer: (pending)
QA Reviewer: (pending)
Doc Reviewer: (pending)

Documentation: n/a

Release Notes: n/a

Details: in order to be able to check the clickable property of a view before it has been set and not get a warning, it should be initialized.  This change sets the clickable attribute on the prototype of LzView.
    

Tests:

git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@4394 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c