Change 20070318-sallen-8 by sallen@sallen-new on 2007-03-18 19:22:10 PDT     in /cygdrive/c/laszlo/svn/openlaszlo/trunk/lps/components/incubator     for http://svn.openlaszlo.org/openlaszlo/trunk/lps/components/incubator Summary: somehow missed from last checkin, tooltip should extend node not view
Bugs Fixed: LPP-3715 tooltip should extend node not view 

Technical Reviewer: pkang@laszlosystems.com
QA Reviewer: (pending) 
Doc Reviewer: (pending) 

Documentation: 

Release Notes: 

Details: 
This change makes it so you can attach a tooltip to a view with a resource 
and not get errors telling you that a view with a resource can't 
have subviews. There's a single tooltipvew that is manipulated by 
the tooltip class -- so tooltip doesn't actually need to be a view. 

git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@4409 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c