Change 20070323-sallen-8 by sallen@sallen-new on 2007-03-23 12:38:56 PDT     in /cygdrive/c/laszlo/svn/openlaszlo/trunk-clean     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: mouse events for input and selectable text

Bugs Fixed: LPP-3753 onmouseup and onmousedown not reported by LzGlobalMouse for inputtext and selectable text

Technical Reviewer: ptw@openlaszlo.org (reviewed code)
QA Reviewer: (pending)
Doc Reviewer: (pending)

Documentation:

Details: in sending events the LFC has special logic to tell if the user clicks on an input text or selectable text field (since we don't get the movie clip that was clicked on from Flash, but we can find out what movie clip is the selection.  This logic is in LzModeManger.  To find the view associated with the movie clip, we rely on a property that the text views set on their corresponding movie clip. That used to be called 'view' but was changed to '__lzview' at some point for inputtext and seems to have gotten lost for test.  This change fixes that.



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@4468 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c