Change 20070923-ben-l by ben@slim.local on 2007-09-23 15:15:48 PDT     in /Users/ben/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Add iPhone-specific hints to lzt=html wrapper to make iPhone apps look nice

New Features: apps are not zoomable or scalable with two-finger scaling on the iPhone

Bugs Fixed:

Technical Reviewer: ptw 
QA Reviewer: (pending)
Doc Reviewer: (pending)

Documentation:
As described in http://weblog.openlaszlo.org/archives/2007/07/our-first-iphone-app/
OpenLaszlo apps can look pretty decent on the iPhone just by adding a single
line to the html wrapper. 

To see it in action, check out and run...
http://svn.openlaszlo.org/labs/newsmatch/
with and without this patch applied, with lzt=html. index.html,
in that directory, has links to the apps with the correct
query args, lzr=dhtml&lzt=html. 

Without this patch, the app is scaled down to too small on the iPhone, 
and the user can zoom and pan. With this patch, it appears at a good 
size and is not zoomable. However, it is still pannable; I'm not sure
how to stop that.  

This meta tag should be ignored by all other browsers. It is
a very small download cost. 

Release Notes:

This change makes it possible to get apps to look right 
on the iPhone without hand-editing wrappers.

Details:
    
Tests:
(see above, with newsmatch-labs)



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@6568 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c