Change 20071109-hqm-6 by hqm@IBM-2E06404CB67 on 2007-11-09 09:23:21 EST     in /cygdrive/c/users/hqm/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: prevent spurious compiler warnings when overriding apply()
method on subclases of <state>

New Features:

Bugs Fixed: LPP-5067, LPP-5065

Technical Reviewer: ptw
QA Reviewer: pbr
Doc Reviewer:

Documentation:

Release Notes:

Details:


I modified the compiler to not warn if the attribute being overriden
was declared explicitly with a value of "true" for it's 'override' property.

So in lfc.lzx, we declare apply as non overridable, on the base Instance

<interface name="Instance" extends="Object">
 <method name="apply" override="false"/>

but "state" declares it overridable, so any subclass of state can override it


<interface name="state" extends="node">
 <attribute name="apply" override="true" />


I made the compiler not only not warn if you override the method with
another method, but not warn if you override an attribute with a
method. That is a little too loose I guess, but it only happens when you
declare "override=true" on an attribute.


Tests:

test/compiler_errors/override-state.lzx
override-builtin.lzx
smokecheck
amazon
calendar
examples/components/style_example.lzx




git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@7200 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c