Change 20071112-ben-6 by ben@slim.local on 2007-11-12 15:09:33 PST     in /Users/ben/src/svn/openlaszlo/trunk/docs/src/developers     for http://svn.openlaszlo.org/openlaszlo/trunk/docs/src/developers
Summary: Move components intro tutorial into tutorials directory

Bugs Fixed: LPP-5005 problem with the generated html for Chapter 6. Introduction to Components

Technical Reviewer: none
Doc Reviewer: sundman (pending)

Documentation:
The "introduction to components" chapter is presented as a "tutorial" in the developer's
guide, but it was stored on disk as a sibling with the non-tutorial chapters. The
XSLT for generating the html treated the chapter as a tutorial, and uses relative
paths to the includes for css and js. Those paths were wrong for this chapter, 
because the html comp-intro.html didn't live in the right place for a tutorial. 
This change moves the file into the correct place for this file, and makes
the devguide's index.dbk point to the moved chapter. 

Also moved all the associated examples into developers/tutorials/programs. 


Tests:
http://localhost:8080/trunk/docs/developers/tutorials/comp-intro.html
Notice that the fonts and colors are right (meaning that the css is in place)
and that the live examples are present and work. (The edit button does not
work; this is noted in many open bugs.) 



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@7221 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c