Change 20071113-ben-F by ben@slim.local on 2007-11-13 17:47:05 PST     in /Users/ben/src/svn/openlaszlo/trunk-for-merging     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Make reference list events and attributes better

New Features:
Technical Reviewer: none 
Doc Reviewer: sundman (pending)

Documentation:
Events appear in reference in Events list, not in the
attributes table. 

Initial arguments are not explicitly listed; instead, the 
associated attribute property is listed as an attribute. 

Reference for LzNode and LzView cleaned up by marking lots
of attributes and some methods "private". 

Get rid of the declaration of LzView.play, because it's 
conflicting with LzView.play(). This is filed as bug
LPP-5098. 

For faster builds, turn off building the charting and 
graphing reference and the cross reference. The c&g 
reference is explicitly excluded from 4.1 reference 
deliverable. The xref is planned to be included as-is, but
since we're not looking at it yet, let's not spend time 
building it. 

Release Notes:

Details:
This checkin is a merge from paperpie of r7216,7229.     

Tests:



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@7247 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c