Change 20071119-ben-k by ben@slim.local on 2007-11-19 14:04:48 PST     in /Users/ben/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Correct js2doc markup for LzText and basecomponent

Technical Reviewer: none 

Documentation:
Several events in LzText needed to be annotated as events, 
using /** @lzxtype event */ 

basecomponent needed the focusable attribute to be
private, because that attribute is declared on LzView. 

Release Notes:

Details:
    

Tests:
ant clean build doc 
http://localhost:8080/trunk/docs/reference/LzText.html
Notice that the only event listed is "ontext" -- this is
the same as in 3.4. 

http://localhost:8080/trunk/docs/reference/lz.basecomponent.html
Note that the focusable attribute is inherited from view.



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@7325 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c