Change 20071126-ben-g by ben@slim.local on 2007-11-26 15:41:28 PST     in /Users/ben/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Fix broken links on docs landing page

Bugs Fixed: LPP-5148 (partial) Fix broken links in docs

Documentation:
We're not shipping the contributor's reference anymore, so, we
shouldn't link to it from the main landing page. 

Recommend the 4.1 reference but link to 3.4 reference 
as a fallback. 

Tests:
http://localhost:8080/trunk/docs/
Notice all links are valid and correct. 



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@7376 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c