Change 20071126-ben-o by ben@slim.local on 2007-11-26 17:12:03 PST     in /Users/ben/src/svn/openlaszlo/trunk-docbuild     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Make doc build only build once on build machines. 

Bugs Fixed: LPP-3270 Refguide is being built twice on build machines 

Documentation:
The doc build takes a lot of time, and doesn't work well on windows,
at the moment, so we want fine control over when exactly to build it. 
This change adds logic so that we won't build the doc if we've
already built it in this invocation of ant, if we're running on 
a windows box, or if we've set skip.doc to true. 

As before, if you just call
$ ant doc
it should build the doc. 

To skip building the doc, pass in -Dskip.doc=true, like this:
$ ant -Dskip.doc=true build

Most importantly, calling ant dist-one should build the docs
exactly once. 

Tests:
Run ant dist-one in the nightly build. Search for docs/reference/index.html
in the output. Before this change, it appeared twice, indicating that the
docs had built twice. After this change, it only appears once, indicating
that the docs only built once. 

Also, verify that the windows build succeeded, and that it did not 
try to build the docs. 




git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@7378 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c