Change 20071127-ben-L by ben@slim.local on 2007-11-27 16:56:09 PST     in /Users/ben/src/svn/openlaszlo/trunk-work-11-27-07-noon     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Continued tweaks to the don't-build-doc-twice-in-nightly-build fix

Bugs Fixed: (again, I hope) LPP-3270 reference builds twice on nightly build

Technical Reviewer: ptw (pending)
QA Reviewer: mkratt (pending)

Documentation:
When calling an ant target from script in an ant file, as I do in the
"doc" target, it is a good idea to be thorough, if one wants the code
to work on a variety of architectures. To accomplish this, we explicitly
request a local variable for the ant tasks we are creating 
programmatically. This apparently keeps the parent build alive long
enough that when the child build completes, the parent build can
continue. Without explicitly saying "var myTask..." the parent
build would terminate as soon as it invokes the child build, 
and we'd never get to finish the parent build. 

As before, the doc should not build if skip.doc is true, if
the doc has already been built, or if the OS is windows. 

Tests:
Run a nightly build. It should work, and the doc should only be 
built once. 
Already verified on linux-builder that we get through building
the doc and on to the next task, checking for tabs. (In my 
environment, checking for tabs failed because I didn't have
the path to python correct, but that's not what we're 
worrying about right now.) 



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@7396 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c