Change 20071209-ben-o by ben@slim.local on 2007-12-09 20:14:20 PST     in /Users/ben/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Don't issue so many unnecessary warnings in the doc build

Technical Reviewer: none

Documentation:

The js2doc build was issuing the same error over and over: 
WARNING: Unexpected number of child nodes in ... 
I changed this to be a "fine" log message, rather than a warning. 
Now if we want to see it, we just set the log level to Level.FINE, 
but we don't have to see it otherwise. 

Similarly, there were lots of messages during the js2doc2dbk step of the
build regarding unknown types:
    Warning! No jstype found for xxxx
and 
    Warning! No class name found for function synopsis: xxx.xxx.xxx

These are now toggled off with several parameters, set in these files:
in utilities.xsl: 
    <xsl:param name="warn.unspecified.lzxtypes" select="0" />
    <xsl:param name="warn.unspecified.jstypes" select="0" />
In synopsis.xsl
    <xsl:param name="warn.classname.not.found" select="0" />  
    
These circumstances are visible in the reference itself as empty
entries in the Type (tag) and Type (js) columns of the attributes
list. The "no class name found" error is visible in the methods
list in the reference guide. Beneath the name of the method,
it's supposed to say className.methodName( params... ) but for the
cases that used to trigger warnings, it does not prepend a 
class name. 

Tests:
ant build doc 
notice that there are hundreds fewer errors than before. 



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@7492 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c