Change 20071212-ben-n by ben@45.1.59.10.in-addr.arpa on 2007-12-12 17:14:41 PST     in /Users/ben/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Enable building contributor's guide on demand

Documentation:
There is some useful tools information in the contributor's guide. 
This change makes it possible to build just the contributor's guide,
and not the contributor's reference, by calling "ant contributors"
in docs/src. 

This does not change the fact that we don't maintain or ship
the contributor's guide; it just had some tasty information in 
it that I wanted to look at, and this was the best way to get
that information out. Someday, we should start shipping this
information again. 

Tests:
cd docs/src
ant contributors
http://localhost:8080/trunk/docs/contributors/index.html



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@7530 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c