Change 20071213-ben-V by ben@slim.local on 2007-12-13 19:36:49 PST     in /Users/ben/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Include a DOCTYPE in the doctools generated content

Bugs Fixed: LPP-5258 doctools should generate html with a DOCTYPE

QA Reviewer: brynn

Documentation:

The html that the doctools was generating did not have a doctype at all. Some
research (thank you, Tucker) revealed that the lack of a doctype was probably
sending IE into quirks mode, which probably explains the many visual errors we
were seeing only in IE.

This change gets saxon to insert a doctype (XHTML 1.0 Transitional) at the
beginning of every html file it generates. This change *should* avoid quirks
mode in everything we care about except for IE6; apparently, this combination
will still evoke quirks mode in IE6.

When this code makes it into a build, the various IE-specific visual errors should be rechecked; hopefully they will be gone! 

HOWEVER: this change seems to cause tons of extra space to be inserted in program listings. That seems like a problem that we'll just have to deal with; getting a right doctype in is the Right Thing To Do. 

 The first lines of every generated html file now look like this: 

<?xml version="1.0" encoding="UTF-8" standalone="no"?>

<!DOCTYPE html
  PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
<html xmlns="http://www.w3.org/1999/xhtml">

...

based on http://hsivonen.iki.fi/doctype/ this doctype/xml declaration combination will provoke standards mode in almost everything:
NS6
    Mozilla 0.6=0.9.4 and Netscape 6.0--6.2.3
Old Moz
    Mozilla 0.9.5 through 1.1 alpha and Mozilla 1.0
Moz & Safari
    Mozilla 1.0.1, Mozilla 1.1 beta and later, Firefox and Netscape 7, Safari v73 aka. 0.9 through Safari v419.3 aka. 2.0.4, Konqueror 3.5
Opera 9
    Opera 9.0-9.10
Opera 7.5
    Opera 7.5-8.02
    
..."Almost Standards" mode in...    
IE 7 & Opera 7.10
    Windows IE 7 and Opera 7.10-7.23
    
...and quirks mode only in... 
IE 6 & Opera 7.0
    Windows IE 6 and Opera 7.0-7.03
Konq 3.2
   
Tests:
Inspect a single html page output by the doctools in both the reference and the
devguide. Notice that it begins with a doctype. Also note that it is w3c valid
XHTML transitional!



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@7539 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c