Change 20071216-ben-s by ben@slim.local on 2007-12-16 15:20:28 PST     in /Users/ben/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Move "building data drive applications" chapter where it belongs

New Features:

Bugs Fixed: (partial) LPP-5039 examples 38.3 and 38.4 broken in Chapter 38. Building Data-Driven Applications

Technical Reviewer: iorio (pending)

Documentation:

The developer's guide chapter, "Building Data Driven Applications", is 
not a tutorial, but the files lived inside the developers/tutorials
directory. This made it impossible for the doctools to find the example
code listings when building the html for that chapter. 

This change moves the docbook chapter and the related example
code out of the tutorials/ directory and into the top-level
docs/src/developers directory. 

The tutorials/ directory is an unnecessary complication that should be done
away with. 

This is not a complete fix for this bug! The applications don't
actually appear and run; I suspect this is because they can't find the
data files they need. I suspect more will be revealed if these build
output warnings are addressed: 

book.resources.generate:
    [mkdir] Created dir: /Users/ben/src/svn/openlaszlo/trunk/docs/developers/programs
     [copy] Copying 331 files to /Users/ben/src/svn/openlaszlo/trunk/docs/developers/programs
    [mkdir] Created dir: /Users/ben/src/svn/openlaszlo/trunk/docs/developers/images
     [copy] Copying 67 files to /Users/ben/src/svn/openlaszlo/trunk/docs/developers/images
     [copy] Copying 97 files to /Users/ben/src/svn/openlaszlo/trunk/docs/developers/images
    [mkdir] Created dir: /Users/ben/src/svn/openlaszlo/trunk/docs/developers/programs/resources
     [copy] Warning: /Users/ben/src/svn/openlaszlo/trunk/docs/src/developers/resources not found.
    [mkdir] Created dir: /Users/ben/src/svn/openlaszlo/trunk/docs/developers/data
     [copy] Warning: /Users/ben/src/svn/openlaszlo/trunk/docs/src/build/developers/data not found.
    [mkdir] Created dir: /Users/ben/src/svn/openlaszlo/trunk/docs/developers/resources
     [copy] Warning: /Users/ben/src/svn/openlaszlo/trunk/docs/src/build/developers/resources not found.
     [copy] Copying 600 files to /Users/ben/src/svn/openlaszlo/trunk/docs/developers


Release Notes:

Details:
    

Tests:
ant clean developers
Note that the output no longer says... 
     [java] Failure reading file:/Users/ben/src/svn/openlaszlo/trunk/docs/src/build/developers/programs/data_app-$1.lzx: /Users/ben/src/svn/openlaszlo/trunk/docs/src/build/developers/programs/data_app-$1.lzx (No such file or directory) 



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@7561 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c