Change 20071217-ben-l by ben@slim.local on 2007-12-17 13:23:38 PST     in /Users/ben/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Set canvas color for the first live example in Laszlo Basics

New Features:

Bugs Fixed: LPP-5221 (partial) canvas bgcolor does not work in live examples in the dguide

Technical Reviewer: iorio (pending) 

Documentation:
To make a live example have a canvas color, specify the bgcolor in the programlisting's
parameter tag: 
   <parameter role="canvas">height: 100, width: 500, bgcolor: 'green'</parameter>


Tests:
Notice that the first example in 
http://localhost:8080/trunk/docs/developers/tutorials/laszlo_basics.html
has a green background; the rest are white. 

Also note that other canvas on that page and on other pages in the dev
guide continue to have white backgrounds. 



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@7581 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c