Change 20071221-ben-a by ben@slim.local on 2007-12-21 15:42:59 PST     in /Users/ben/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Improvements to contributor-oriented chapters in devguide

Documentation:
Modified the "Source Build" chapter to be up to date for the
4.1.x build, with no python and no jython required. Updated
it with some of the more recent changes on the wiki. This chapter
is now just about as good as 
http://wiki.openlaszlo.org/SubversionBuildInstructions. 

Gave each chapter an id, so the html chunks have nice filenames
instead of just "ch55.html"

Removed the completely empty chapter on release management.



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@7661 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c