Change 20071221-ben-X by ben@slim.local on 2007-12-21 21:24:33 PST     in /Users/ben/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Add explanation of first part of reference doc build

Documentation:
This checkin fills in the documentation toolchain chapter in the 
developer's guide with an overview and then with a walkthrough of
each of the code entry points into the reference build data flow,
with examples of the relevant source and intermediate products. 

I've also added two diagrams based on the doc-toolchain.graffle
dataflow diagram; the first, doc_overview_simple.png, is a 
much simpler (but less accurate) representation of the doctools. 
The second, doc_reference_details.png, is a simplified magnification 
of three strands of the reference guide workflow. 

We're keeping the original, highly-detailed data flow diagram,
of course. 

Release Notes:

Details:
    

Tests:



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@7663 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c