Change 20071226-ben-7 by ben@slim.local on 2007-12-26 18:06:19 PST     in /Users/ben/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Checkpointing more work on doctools explanatory chapter

Documentation:

Filled out a section about how js2doc intermediate files are
transformed into docbook files. This is the second major phase
of the documentation toolchain. (The remaining major phase is 
docbook to html.) 

A new diagram helps focus attention on this phase of the transformation.

This checkin also includes some advice on how to learn and
modify the documentation toolchain.

This checkin also adds emphasis to fragments of the program
listings. 

Tests:



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@7671 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c