Change 20080106-lou-G by lou@loumac.local on 2008-01-06 12:50:41 AST     in /Users/lou/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: document how the copyright notice gets included in the dguide and reference

Bugs Fixed: LPP-5329

Technical Reviewer: (pending)
QA Reviewer: (pending)
Doc Reviewer: (pending)

Details: added description to the doc toolchain chapter of the dguide.

5.1. How the copyright gets generated

There is a copyright notice at the bottom of every dguide and reference page. This notice is generated during the docbook transform by the template user.footer.content in the file common-html.xsl. To change to copyright date for all generated documents, change the date in this template.

The copyright notice for files that are not generated is entered manually. 
    
Tests:



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@7745 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c