Change 20080114-lou-Z by lou@loumac.local on 2008-01-14 11:52:34 AST     in /Users/lou/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: text was wrapping around example window badly.

Bugs Fixed: LPP-5079

Technical Reviewer: (pending)
QA Reviewer: (pending)
Doc Reviewer: (pending)

Details: This is not really a bug, but a function of how wide the browser window is.
If you make the browser window wide enough, the complete text appears to the right
of the example window as designed.

I did reduce the space between the example window and the text, so this is less likely
to happen.
    

Tests: visual check



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@7822 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c