Change 20080117-lou-W by lou@loumac.local on 2008-01-17 08:57:23 AST     in /Users/lou/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: fix link to the style reference in Laszlo Explorer.

Bugs Fixed: LPP-4322

Technical Reviewer: (pending)
QA Reviewer: (pending)
Doc Reviewer: (pending)

Details: Change the link so it goes directly to the lz.style.html page.
Not that this is not an ideal fix. What should really happen is the link
should take you to the reference, with the navigation pane on the left and
the style page on the right. Because the navigation is actually an lzx
application, I don't think this is possible, so I did the next best thing.
    

Tests: test link



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@7846 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c