Change 20080118-lou-J by lou@loumac.local on 2008-01-18 08:43:18 AST     in /Users/lou/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: fix live example order in the databinding chapter of the dguide

Bugs Fixed: LPP-3936

Technical Reviewer: (pending)
QA Reviewer: (pending)
Doc Reviewer: (pending)

Details: At some point, an example was inserted in this chapter, which threw
off the order of all following examples. Fix the order of the examples so
they appear in the proper sections.

One example file contained the wrong example; revert to the 3.4 version.
    

Tests: check that the examples match the 3.4 doc



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@7851 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c