Change 20080126-bargull-6 by bargull@dell--p4--2-53 on 2008-01-26 16:43:10     in /home/Admin/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Use delegate-id to identify delegate in LzTimer

New Features:

Bugs Fixed: LPP-5379 - "LzTimer never stops calling delegate if called object's toString() value changes"

Technical Reviewer: ptw
QA Reviewer: pbr
Doc Reviewer: (pending)

Documentation:

Release Notes:

Details:
As described by the bug-reporter, if the return-value of the context's "toString"-method changed, the delegate won't be removed from LzTimer. This happens because a delegate itself is used as the key for the LzTimer's "timerList", or better the string-representation of the delegate -> which is the return value of the delegate's "toString"-method. But the "toString"-method also depends on the context's "toString"-method. And I guess I don't need to underline that this may cause several errors for components like <text>...
However, to fix this bug, the delegates are now identified by their delegate-id in the "timerList".


Tests:
see bugreport



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@7947 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c