Change 20080227-lou-m by lou@loumac.local on 2008-02-27 10:27:54 AST     in /Users/lou/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Document that basetree needs a relative, not absolute datapath

Bugs Fixed: LPP-4603

Technical Reviewer: henry
QA Reviewer: (pending)
Doc Reviewer: (pending)

Details:
This was already documented in the reference page for <tree>, but the
markup was wrong, so the example code did not show up. Also, the
code was wrong (the nested <tree> was missing the /).

Fixed markup and example, and duplicated doc in the <basetree> page,
which is what the bug was reported against.

Tests: visual verify



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@8113 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c