Change 20080229-maxcarlson-S by maxcarlson@Roboto on 2008-02-29 15:42:21 PST     in /Users/maxcarlson/openlaszlo/trunk-clean     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Fix html frame support with SSL.

New Features:

Bugs Fixed: LPP-5532: Iframe in iframemanager.js causes unsecure items warning in IE with SSL turned on

Technical Reviewer: promanik
QA Reviewer: lhenrywilkins@laszlosystems.com
Doc Reviewer: (pending)

Documentation:

Release Notes:

Details: Sending on behalf of Lorien with blank.html deleted from svn, and a minor typo fixed.

Need to initialize iframes that are used in the history mechanism and html tag to a value that will not be flagged as an unsecure item when the app is running with SSL.  When no initial source value is set on an iframe it defaults to about:blank, which in IE resolves to http://about:blank even if the container page protocol is https.

Setting the initial source to javascript:"" solves the problem.
This change adds this initialization to the iframemanager, and changes the mechanism used in lzhistory
to be the one described above as it is a simpler solution. Previously in lzhistory.js the iframe source 
was initialized to blank.html, which required a page called blank.html be available when the script
executes.
    

Tests: This can be tested by running the test case at lps/components/extensions/test/html-swf.jsp through
SSL in IE 6.0. Before this fix in in place you will get an unsecure content warning when the page loads.



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@8146 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c