Change 20080328-hqm-N by hqm@badtzmaru.local on 2008-03-28 14:42:10 EDT     in /Users/hqm/openlaszlo/trunk4     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: emit override on class methods that need it

New Features:

Bugs Fixed:

Technical Reviewer: (pending)
QA Reviewer: (pending)
Doc Reviewer: (pending)

Documentation:

Release Notes:

Details:

When adding a method, NodeModel checks if the superclass has a method,
and if so adds "override" to the declaration. 

Note, causes a compiler parser error when the method is not a real
class method, such as a method on an instance, because the "override"
keyword cannot be used for some function closure expression, so we
won't emit 'override' or "final" modifiers for non-class-methods for
now.


Tests:

test/compiler_errors/class-with-method-override.lzx   runs in swf9

demos/amazon compiles and runs without error in swf8 and dhtml
demos/calendar compiles and runs without error in swf8 and dhtml
smokecheck  compiles and runs without error in swf8 and dhtml
examples/components/style_example ""




git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@8465 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c