Change 20080403-lou-N by lou@loumac.local on 2008-04-03 05:23:31 AST     in /Users/lou/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: copy release-notes.css from pagan-deities to trunk

Bugs Fixed: LPP-5192 (partial)

Technical Reviewer: (pending)
QA Reviewer: (pending)
Doc Reviewer: (pending)

Details: this style sheet was created for the pagan-deities release to display the bug table properly. It was not needed for devildog, but it will be needed for ringding.

release-notes.html still needs to be changed to use this style sheet.
    

Tests:



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@8533 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c