Change 20080411-hqm-4 by hqm@badtzmaru.home on 2008-04-11 15:22:41 EDT     in /Users/hqm/openlaszlo/trunk5     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary:  fix for XML RPC, and some other brain damage in data loader

New Features:

Bugs Fixed: LPP-5481

Technical Reviewer: max
QA Reviewer: pbr
Doc Reviewer: (pending)

Documentation:

Release Notes:

Details:

+ A while back we discovered that that data loading API would not let
you POST data to a url while also supporting an independent query
string in the URL.

So I had made some fixes for that, both in SOLO and proxied mode.  In
proxied mode I moved the place where the POST payload was stored out
of the "url" arg.

That caused some trouble because there is older code in
XMLRPCDataSource.java which expects to be able to stick
"lzpostbody=DATA" into the url and call the
HTTPDataSource.getHTTPData() method, which no longer expects to see
that data there.

I fixed this by putting back the code to look for the "lzpostbody" arg
inside the URL as well as in the proxy request parameters.

Sorry this is so baroque. The main idea going forward is that we have
the LzHTTPLoader API which looks like the javascript XMLHTTPRequest.
Everything in LzHTTPDataProvider is geared to talk to that API; a URL
with query args, and an optional POST content string.

The SWF kernel loader code has to then unpack that and make it work
with the brain-dead Flash LoadVars HTTP API as much as possible, in
solo and proxied mode.


Tests:

test/lfc/data/alldata.lzx, in dhtml and swf
demos/amazon
XMLRPC test case in bug report



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@8635 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c