Change 20080428-Philip-4 by Philip@Philip-DC on 2008-04-28 20:25:04 EDT      in /cygdrive/f/laszlo/svn/src/svn/openlaszlo/trunk      for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Fix tree_example in dhtml

New Features:

Bugs Fixed: LPP-5748

Technical Reviewer: ptw
QA Reviewer: (pending)
Doc Reviewer: (pending)

Documentation:

Release Notes:

Details:
basetree.lzx was trying to create objects using the string name of the class. I also changed the calls to setVisible() to remove the deprecated warnings.

This is Tucker's definitive explanation of the issue:

The correct fix is to fix getChildClass, which currently says:

          <method name="getChildClass">
              if (this.isleaf) return null;
              return this.constructor.classname;
          </method>

I guess the confusion is that it implies it returns a class, but it is actually returning a class _name_ (a string).

There are two choices of how to fix it.  If you need a tag name, because you are going to look it up in the `lz` module (as is done in say, baselist#addItem) you want:

    this.constructor.tagname

If you look this up in lz:

    lz[this.constructor.tagname]

you will get the constructor for the class that implements that tag.

But, if you just want the constructor directly, just say:

    this.constructor

It should always be the case that:

    lz[this.constructor.tagname] === this.constuctor

There's a bunch of confusion that has been introduced because we have made classnames non-global, and they don't necessarily match the name of the tag they implement.  (They never did for the LFC, we are now just uniform for all classes that implement tags).

I guess the confusion is increased because to the LZX programmer, when they say:

    <class name="foo" ...

they think they have created a class named `foo`, but really they have created a (Javascript) class with an anonymous name that implements the tag named `foo`.  The mapping from tag name to implementation class is through the `lz` table.  Perhaps we should rename the Javascript `classname` property to `implementationclassname` or perhaps get rid of it altogether, as it seems to only be a source of confusion...  I'll leave it to you if you want to file an improvement request to that effect.



Tests:
tree example runs in dhtml:
http://localhost:8080/trunk/examples/components/tree_example.lzx?lzr=dhtml

Files:
M      lps/components/base/basetree.lzx

Changeset: http://svn.openlaszlo.org/openlaszlo/patches/20080428-Philip-4.tar



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@8931 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c